Bug: turn on SD_FAST_XFER_AKTIV and compilation fails if PID is enabled.
Fix:
If SD_FAST_XFER_AKTIV is defined, fast_xfer() only declares
g_heater_pwm_val if PIDTEMP is defined, but later assigns
g_heater_pwm_val=0 regardless whether PIDTEMP is defined or not. If
PIDTEMP is undefined, g_heater_pwm_val is undeclared at compile time and
generates the appropriate error. I solved by not using g_heater_pwm_val
if PIDTEMP is undefined (and therefor undeclared.)

Also, the extern declaration was as an int, when g_heater_pwm_val is
defined as a volatile unsigned char. This caused a duplicate declaration
error, which was solved by changing the declaration in fast_xfer() to
match the definition. I didn't dig deeper to try to understand if the
mismatch was intentional or test operation other than verify the compile
errors no longer occur.